feat: Setup half of location posting view
Sets up the first part of the location posting view:
a screen to allow the user to enter their location as a string.

Next step involves converting the string to a CLLocation, adding
a UIMapView, and making some UI tweaks to move ahead to actually
posting the user's location to Parse.